+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
-## HEAD
+##### v0.1.0, 2015-02-28
+- Add default build tasks
 
-## v0.1.0, 2015-02-28 - Add default build tasks
+##### v1.0.0, 2015-05-05
+- Refactor Gruntfile.main.js und config.default.js
+- Delete nicht gebrauchte tasks (karma, autoprefixer, ...)
+- Add README.md